refactor(userspace-dp): extract helpers/session_sync module
Begin the #6234 decomposition of server/helpers.rs, a 1.4k-line
cold-path grab-bag that had become an ownership boundary for five
distinct subsystems (status projection, session-sync reconstruction,
binding/queue planning, and state persistence) behind a single
`use super::super::*` glob.

Convert server/helpers.rs into a server/helpers/ directory module and
move the HA session-sync reconstruction — build_synced_session_key,
build_nat64_reverse_rebuild, build_synced_session_entry, and
parse_session_sync_mac — into helpers/session_sync.rs with explicit
per-module imports (crate::afxdp, crate::protocol::SessionSyncRequest)
in place of the crate-root glob.

Pure code-motion: function bodies are byte-for-byte identical. mod.rs
re-exports the submodule (`pub(crate) use session_sync::*`) so every
call site keeps reaching the helpers through the same
`server::helpers::<name>` path (handlers, main.rs, tests). The remaining
inline helpers keep the original glob until the follow-on commits peel
them into planning/persistence/status siblings.

Validation: `cargo build` on userspace-dp is green.

Refs #6234
